refactor(backend-service-user): move AppModule
